+++ tests/clinq.Tests.cpp
@@ -1,0 +1,10 @@
+/// @file CLinq.Tests.cpp
+/// Contains tests for the CLinq project.
+
+#define CATCH_CONFIG_MAIN
+#include "catch.hpp"
+
+namespace CLinq
+{
+
+}