+++ CHANGELOG.md
@@ -4,3 +4,5 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 🔖 [0.1.0] - Unreleased
+### ✨ Added
+- Unit test project using Catch2.